add rubocop to the project, run and fix any obvious offenses
the earliest version of ruby that I could get rubocop working with is
2.4 - earlier than that would mean setting the docker image to an
earlier version of ruby, which would introduce other problems with the
development tooling.

Unfortunately, the earliest version of ruby we actually want to
support is 2.0, so this could be a problem. Howwever, the CI will run
the test suite on 2.0 and 2.3, so we should get a warning if rubocop
tries to make is write code which is incompatible with the target versions.